+++ crates/core/src/config/mod.rs
@@ -275,6 +275,7 @@ pub enum Language {
     French,
     Hungarian,
     Portuguese,
+    Polish,
     Russian,
     Slovak,
     Swedish,
@@ -297,6 +298,7 @@ impl std::fmt::Display for Language {
                 Language::Hungarian => "Magyar",
                 Language::Norwegian => "Norsk Bokmål",
                 Language::Portuguese => "Português",
+                Language::Polish => "Polski",
                 Language::Russian => "Pусский",
                 Language::Slovak => "Slovenčina",
                 Language::Spanish => "Español",
@@ -310,7 +312,7 @@ impl std::fmt::Display for Language {
 
 impl Language {
     // Alphabetically sorted based on their local name (@see `impl Display`).
-    pub const ALL: [Language; 14] = [
+    pub const ALL: [Language; 15] = [
         Language::Czech,
         Language::Danish,
         Language::German,
@@ -320,6 +322,7 @@ impl Language {
         Language::Hungarian,
         Language::Norwegian,
         Language::Portuguese,
+        Language::Polish,
         Language::Russian,
         Language::Slovak,
         Language::Swedish,
@@ -342,6 +345,7 @@ impl Language {
             Language::Slovak => "sk_SK",
             Language::Turkish => "tr_TR",
             Language::Portuguese => "pt_PT",
+            Language::Polish => "pl_PL",
             Language::Ukrainian => "uk_UA",
         }
     }