+++ src/localization.rs
@@ -32,6 +32,8 @@ pub static LOCALIZATION_CTX: Lazy<JSONGetText<'static>> = Lazy::new(|| {
         "locale/tr.json",
         "pt_PT",
         "locale/pt.json",
+        "pl_PL",
+        "locale/pl.json",
         "uk_UA",
         "locale/uk.json"
     )