+++ CHANGELOG.md
@@ -16,6 +16,8 @@ and `Removed`.
 
 ### Added
 
+- New language added to Ajour:
+  - Polish (thanks KasprzykM)
 - Ability to select Screenshots when backing up.
 
 ## [1.2.0] - 2021-05-18